fix: cluster bug

+++ tests/test_return_probs.py
@@ -233,7 +233,7 @@ def test_top_k_top_p_sampling_acc(batch_size=1):
     print(f'{diff_golden=}')
 
 if __name__ == "__main__":
-    torch.set_printoptions(threshold=float('inf'), precision=6)
+    torch.set_printoptions(threshold=float('inf'), precision=32)
     test_top_k_top_p_sampling_acc(1)
 
     # test_top_k_top_p_sampling_performance(1)
